pyproject.toml
[project]
name = "cse447-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "kenlm",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "wandb>=0.25.0",
]

[dependency-groups]
dev = [
    "datasets==3.6.0",
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[tool.uv.sources]
kenlm = { git = "https://github.com/kpu/kenlm.git" }
